Create file_utils python file
