+++ minis/downloader/__init__.py
@@ -1,3 +1,4 @@
 from .downloader import *
 from .pages import *
 from .filters import *
+from .file_utils import *
