+++ minis/downloader/file_utils.py
@@ -1,0 +1,14 @@
+import os
+
+def createDir(path):
+    try:
+        os.mkdir(path)
+    except OSError:
+        print ("Creation of the directory %s failed" % path)
+    else:
+        print ("Successfully created the directory %s " % path)
+
+def writeToFile(html, fileName):
+    with open(fileName, 'w') as f:
+        f.write(html)
+
